- CharacterObject is now a gameobject - CharacterObject vector is now removed, characterobjects are now placed in gameobject vector - fix fov not resetting upon level change, allowing fov glitch when switching level while aiming down sights - change some referenced objects to const, also some objects to pass by reference - Testmode will now affect the ai and camera debug rendering - fix most warnings

+++ Life/Source/Enemy.cpp
@@ -28,9 +28,9 @@ void Enemy::update(double dt, float speed)
 {
 	if (this->getMove())
 	{ 
-		this->acc = this->force * (double)(1 / this->mass);
+		this->acc = this->force * (float)(1 / this->mass);
 		this->velo += acc;
-		this->getBound()->moveTo(this->getBound()->getOrigin() + (velo * speed * dt));
+		this->getBound()->moveTo(this->getBound()->getOrigin() + (velo * speed * static_cast<float>(dt)));
 	}
 
 }
