+++ Life/Source/aPlayer.cpp
@@ -28,11 +28,12 @@ void aPlayer::update(double dt, float speed)
 {
 	if (this->getMove())
 	{
-		this->acc = this->force * (double)(1 / this->mass);
+		this->acc = this->force * (float)(1 / this->mass);
 		this->velo += acc;
 		this->getBound()->moveTo(this->getBound()->getOrigin() + (velo * speed * static_cast<float>(dt)));
 	}
 }
+
 /******************************************************************************/
 /*!
 \brief
